Контроллер для скачивания файлов
